# Linux fixes. Disabled CONFIG_USE_MMGR by default, since it seems unnecessary and it makes debug builds slow (extremely slow under Valgrind) due to symbol lookups when allocating memory. Fixed minor warnings and errors.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@4927 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation/Simulation.cpp
@@ -136,7 +136,7 @@ void CSimulation::Interpolate(double frameTime, double offset)
 
 void CSimulation::Simulate()
 {
-	float time = m_pTurnManager->GetTurnLength();
+	uint time = m_pTurnManager->GetTurnLength();
 
 	PROFILE_START( "scheduler tick" );
 	g_Scheduler.update(time);
